pyproject.toml
[project]
name = "seneca"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "kafka-python>=2.2.15",
    "orjson>=3.11.3",
    "prometheus-client>=0.22.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.5",
    "spacy>=3.8.7",
    "tenacity>=9.1.2",
    "websockets>=15.0.1",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
]
